Added initial version of Furl
This version is the result of a couple of months' of prototyping and
tweaking. It supports exactly the feature set I've needed so far, which
is GET and POST requests, and being able to specify the Content-Type
header for a request.

Lots of other scenarios (such as PUT and DELETE) requests can be added
later, but so far, I haven't needed them.